+++ CHANGELOG.md
@@ -6,6 +6,7 @@
      documents.
    - Change "3_months" graphs to "6_months" graphs in all documents
      containing history objects.
+   - Remove the "fingerprint" parameter.
 
 
 # Changes in version 6.2-1.17.1 - 2018-08-17